the big lot of the changes
the lot of em

+++ code/game/objects/items/colorizers/colorizer.dm
@@ -13,6 +13,14 @@
     var/apply_item_state = null
     var/apply_righthand_file = null
     var/apply_lefthand_file = null
+    var/delete_me = TRUE
+
+/obj/item/colorizer/examine(mob/user)
+	. = ..()
+	if(uses_left)
+		. += "It has [uses_left] use\s left."
+	else
+		. += "It is empty."
 
 /obj/item/colorizer/attack_self(mob/user)
     var/obj/item/target_atom = user.get_inactive_held_item()
@@ -31,6 +39,9 @@
 
 /obj/item/colorizer/proc/do_colorize(atom/to_be_colored, mob/user)
     if(!to_be_colored)
+        return
+    if(uses_left == 0 && delete_me == FALSE)
+        to_chat(user, "<span class='warning'>This colorizer is empty!</span>")
         return
     if(!is_type_in_list(to_be_colored, allowed_targets) || is_type_in_list(to_be_colored, forbidden_targets))
         to_chat(user, "<span class='warning'>This colorizer is not compatible with that!</span>")
@@ -53,5 +64,5 @@
     to_chat(user, "<span class='notice'>Color applied!</span>")
     playsound(src, 'sound/effects/spray.ogg', 5, TRUE, 5)
     uses_left --
-    if(!uses_left)
+    if(!uses_left && delete_me == TRUE)
         qdel(src)
